Refactor struttura: * Riorganizzata la struttura in cartelle * Cambiato l'authorization header in un custom header per compatibilità con safari
