+++ src/components/common/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
-import Authentication from '../lib/Authentication';
+import Authentication from '../../lib/Authentication';
 
 const PrivateRoute = ({ component: Component, ...rest }) => (
     <Route
