+++ src/components/news/NewsCard.js
