+++ src/components/study-groups/StudyGroupCard.js
@@ -3,11 +3,11 @@ import React, { Component } from 'react';
 import Moment from 'react-moment';
 import 'moment/locale/it';
 
-import history from '../lib/history';
+import history from '../../lib/history';
 
-import LocalStorage from '../lib/LocalStorage';
+import LocalStorage from '../../lib/LocalStorage';
 
-import { getInitials, getCourseColor } from '../lib/Utils';
+import { getInitials, getCourseColor } from '../../lib/Utils';
 
 import {
     withStyles,
