+++ src/containers/StudyGroupContainer.js
@@ -7,8 +7,8 @@ import {
     CircularProgress
 } from '@material-ui/core';
 
-import StudyGroup from '../components/StudyGroup';
-import { showNotifier } from '../components/Notifier';
+import StudyGroup from '../components/study-groups/StudyGroup';
+import { showNotifier } from '../components/common/Notifier';
 
 class StudyGroupContainer extends Component {
     state = {
